requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.15.1
graphviz==0.20.3
pydot==2.0.0
pyparsing==3.1.2
sqlparse==0.5.0
